requirements.txt
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
Deprecated==1.2.6
dj-database-url==0.5.0
Django==2.1.7
django-rest-auth==0.9.5
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
ecdsa==0.13.2
future==0.17.1
gunicorn==19.9.0
idna==2.8
imageio==2.5.0
kiwisolver==1.1.0
matplotlib==3.1.1
networkx==2.3
numpy==1.17.2
objectpath==0.6.1
Pillow==6.1.0
pyasn1==0.4.6
pycodestyle==2.5.0
PyGithub==1.43.8
PyJWT==1.7.1
pyparsing==2.4.2
python-dateutil==2.8.0
python-decouple==3.1
python-jose==3.0.1
pytz==2018.9
PyWavelets==1.0.3
requests==2.21.0
rsa==4.0
scikit-image==0.15.0
scipy==1.3.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
wrapt==1.11.2
